:recycle: Update code that was not properly merged when doing rebasing
