+++ src/openpersonen/contrib/demo/converters/partner.py
@@ -1,12 +1,12 @@
 from django.conf import settings
 
-from openpersonen.contrib.utils import is_valid_date_format
 from openpersonen.features.country_code_and_omschrijving.models import (
     CountryCodeAndOmschrijving,
 )
 from openpersonen.features.gemeente_code_and_omschrijving.models import (
     GemeenteCodeAndOmschrijving,
 )
+from openpersonen.utils.helpers import is_valid_date_format
 
 
 def convert_partner_instance_to_dict(partner):
